\name{matrixmarket-package}
\alias{matrixmarket-package}
\alias{matrixmarket}
\docType{package}
\title{
\packageTitle{matrixmarket}
}
\description{
\packageDescription{matrixmarket}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{matrixmarket}
\packageIndices{matrixmarket}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{matrixmarket}

Maintainer: \packageMaintainer{matrixmarket}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
